@@ -9,69 +9,68 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "watchMono": "ng build --watch --configuration=debugMono --output-path=../Overseer.Daemon.Mono/bin/Debug/OverseerUI",
-    "buildDebugMono": "ng build --configuration=debugMono --output-path=../Overseer.Daemon.Mono/bin/Debug/OverseerUI",
-    "buildReleaseMono": "ng build --prod --configuration=productionMono --output-path=../Overseer.Daemon.Mono/bin/Release/OverseerUI",
-    "watchStatic": "ng build --watch --configuration=debugStatic --output-path=./dist-static",
-    "buildDebugStatic": "ng build --configuration=debugStatic --output-path=./dist-static",
-    "buildReleaseStatic": "ng build --configuration=productionStatic --output-path=./dist-static",
+    "watchArmv6": "ng build --configuration=debugArmv6 --output-path=../Overseer.Daemon.Mono/bin/Debug/OverseerUI --watch",
+    "bulldArmv6": "ng build --prod --configuration=productionArmv6 --output-path=../Overseer.Daemon.Mono/bin/Release/OverseerUI",
+    "buildArmv6:debug": "ng build --configuration=debugArmv6 --output-path=../Overseer.Daemon.Mono/bin/Debug/OverseerUI",
+    "watchLocal": "ng build --watch --configuration=debugLocal --output-path=./dist-static",
+    "buildLocal": "ng build --configuration=productionLocal --output-path=./dist-static",
+    "buildLocal:debug": "ng build --configuration=debugLocal --output-path=./dist-static",
     "setVersion": "npm version"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.15",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@aspnet/signalr": "^1.1.4",
     "@ngx-progressbar/core": "^5.3.2",
-    "@ngx-translate/core": "^10.0.2",
-    "@ngx-translate/http-loader": "^3.0.1",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
     "@types/bcryptjs": "^2.4.2",
     "@types/ua-parser-js": "^0.7.33",
     "@types/url-parse": "^1.4.3",
-    "@zerohouse/router-tab": "0.0.7",
     "bcryptjs": "^2.4.3",
-    "core-js": "^2.6.9",
+    "core-js": "^3.4.1",
     "hammerjs": "^2.0.8",
     "jquery": "^3.4.1",
-    "ngx-indexed-db": "^2.2.0",
+    "ngx-indexed-db": "^3.0.5",
     "ngx-logger": "^4.0.7",
     "ngx-store": "^2.1.0",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.3",
     "signalr": "^2.4.1",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "ua-parser-js": "^0.7.20",
     "url-parse": "^1.4.7",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.4",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/jasmine": "^2.8.16",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "^0.803.19",
+    "@angular/cli": "^8.3.19",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.4.6",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "~12.12.8",
+    "codelyzer": "^5.2.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.1.6"
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.5.2",
+    "tslint": "~5.20.1",
+    "typescript": "^3.5.3"
   }
 }
