@@ -1,40 +1,43 @@
 import { Injectable } from "@angular/core";
 import { Machine } from "../../../models/machine.model";
-import { machineStoreName, openDatabase } from "./open-database-function";
+import { NgxIndexedDBService } from "ngx-indexed-db";
 
 @Injectable({ providedIn: "root" })
 export class MachineStorageService {
+    constructor(private db: NgxIndexedDBService) {}
+
+    executeDbRequest<T>(action: () => T): T {
+        this.db.currentStore = "machines";
+        return action();
+    }
 
     async createMachine(machine: Machine) {
-        const db = await openDatabase();
-        return db.add(machineStoreName, machine);
+        machine.id = await this.executeDbRequest(() => this.db.add(machine));
+        return machine;
     }
 
     async getMachines(): Promise<Machine[]> {
-        const db = await openDatabase();
-        return db.getAll(machineStoreName);
+        return await this.executeDbRequest(() => this.db.getAll());
     }
 
     async getMachineById(machineId: number): Promise<Machine> {
-        const db = await openDatabase();
-        return db.getByKey(machineStoreName, machineId);
+        return await this.executeDbRequest(() => this.db.getByID(machineId));
     }
 
     async updateMachine(machine: Machine) {
-        const db = await openDatabase();
-        return db.update(machineStoreName, machine);
+        return await this.executeDbRequest(() => this.db.update(machine, machine.id));
     }
 
     async deleteMachine(machineId: number) {
-        const db = await openDatabase();
-        return db.delete(machineStoreName, machineId);
+        return await this.executeDbRequest(() => this.db.deleteRecord(machineId));
     }
 
     async updateMachines(machines: Machine[]) {
-        const db = await openDatabase();
-        const promises: Promise<any>[] = [];
-        machines.forEach(machine => promises.push(db.update(machineStoreName, machine)));
-
-        return Promise.all(promises);
+        const self = this;
+        return await this.executeDbRequest(() => {
+            const promises: Promise<any>[] = [];
+            machines.forEach(machine => promises.push(self.db.update(machine, machine.id)));
+            return Promise.all(promises);
+        });
     }
 }
