@@ -1,5 +1,5 @@
 export const environment = {
     production: false,
-    dbVersion: 5,
+    dbVersion: 6,
     appVersion: "1.0.15"
 };
